refactor RatingFactor
